# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users (admin only)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all users"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user by ID"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/progress/weight": {
      "post": {
        "operationId": "ProgressController_createWeightLog",
        "summary": "Log a new weight",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeightLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Weight logged successfully"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ProgressController_getWeightLogs",
        "summary": "Get all weight logs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all weight logs"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/progress/weight/range": {
      "get": {
        "operationId": "ProgressController_getWeightByDateRange",
        "summary": "Get weight logs by date range",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "description": "Start date in ISO format",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "description": "End date in ISO format",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return weight logs in date range"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/progress/weight/{id}": {
      "get": {
        "operationId": "ProgressController_getWeightLog",
        "summary": "Get weight log by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return weight log by ID"
          },
          "404": {
            "description": "Weight log not found"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProgressController_updateWeightLog",
        "summary": "Update weight log",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeightLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Weight log updated successfully"
          },
          "404": {
            "description": "Weight log not found"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ProgressController_removeWeightLog",
        "summary": "Delete weight log",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Weight log deleted successfully"
          },
          "404": {
            "description": "Weight log not found"
          }
        },
        "tags": [
          "progress"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Fitness Tracker API",
    "description": "API for fitness and nutrition tracking application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Password123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          }
        }
      },
...
      "CreateWeightLogDto": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "example": 75.5
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T08:00:00Z"
          }
        },
        "required": [
          "weight"
        ]
      },
      "UpdateWeightLogDto": {
        "type": "object",
        "properties": {
          "weight": {
            "type": "number",
            "example": 75.5
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-10-15T08:00:00Z"
          }
        }
      }
    }
  }
}